[Beta] Refactor object designer layout
