+++ src/renderer/app/views/object_designer/designer.style.scss
@@ -13,7 +13,7 @@
 }
 
 :host ::ng-deep .ant-card-small > .ant-card-body {
-  padding: 0;
+  padding: 2px;
 }
 
 :host ::ng-deep .ant-divider-horizontal {
@@ -28,7 +28,13 @@
   background: #ffffff;
 }
 
-:host ::ng-deep .ant-tree {
-  border-radius: 0px;
-  border-right: 1px solid rgba(0,0,0,.06);
+:host ::ng-deep .custom-select {
+  width: 200px;
+  border-right: #f5f6f5 1px solid;
+  margin-right: 10px;
 }
+
+//:host ::ng-deep .ant-tree {
+//  border-radius: 0px;
+//  border-right: 1px solid rgba(0,0,0,.06);
+//}
