+++ src/renderer/app/views/object_designer/layout/content/components/create_table/create.table.style.scss
@@ -1,0 +1,3 @@
+:host ::ng-deep .action-button-success {
+  color: #00c853;
+}
