+++ src/renderer/app/views/object_designer/layout/content/components/preview/database/detail.database.style.scss
@@ -17,3 +17,7 @@
 :host ::ng-deep .ant-divider-horizontal {
   margin: 0;
 }
+
+:host ::ng-deep .ant-card-small>.ant-card-head>.ant-card-head-wrapper>.ant-card-head-title {
+  padding: 8px 10px;
+}
