+++ src/renderer/app/views/object_designer/layout/content/components/preview/table/detail.table.component.ts
