+++ src/renderer/app/views/object_designer/layout/content/layout.content.component.ts
@@ -2,7 +2,6 @@ import { AfterViewInit, Component, EventEmitter, Input, Output } from "@angular/
 import { DesignerApplyData } from "@renderer/app/views/object_designer/model/designer.apply.data";
 import { TypeEnum } from "@renderer/enum/type.enum";
 import { DesignerColumn } from "@renderer/app/views/object_designer/model/designer.column";
-import { AssertUtils } from "@renderer/app/views/object_designer/utils/assert.utils";
 
 @Component({
   selector: 'object-designer-layout-content',
@@ -24,18 +23,5 @@ export class LayoutContentComponent implements AfterViewInit {
   }
 
   ngAfterViewInit(): void {
-    this.handlerInitialize();
-  }
-
-  handlerInitialize() {
-  }
-
-  isCreateTable(): boolean {
-    return AssertUtils.isCreateTable(this.applyData);
-  }
-
-  handlerForward(event: any[]) {
-    this.applyColumns = event;
-    this.emitter.emit(this.applyColumns);
   }
 }
