+++ src/renderer/app/views/object_designer/layout/content/layout.content.view.html
@@ -1,18 +1,8 @@
-<ng-container *ngIf="(applyData.isOpen && !isCreateTable()); else createTable">
-  <ng-container *ngIf="applyData.type === actionType.database">
-    <object-designer-layout-content-detail-database [applyData]="applyData">
-    </object-designer-layout-content-detail-database>
-  </ng-container>
-  <ng-container *ngIf="applyData.type === actionType.table">
-    <object-designer-layout-content-detail-table [applyData]="applyData">
-    </object-designer-layout-content-detail-table>
-  </ng-container>
+<ng-container *ngIf="applyData.type === actionType.database">
+  <object-designer-layout-content-detail-database [applyData]="applyData">
+  </object-designer-layout-content-detail-database>
 </ng-container>
-
-<ng-template #createTable>
-  <object-designer-layout-content-operator-table-create *ngIf="isCreateTable()" [applyData]="applyData"
-                                                        [applyColumns]="applyColumns"
-                                                        (emitter)="handlerForward($event)">
-  </object-designer-layout-content-operator-table-create>
-</ng-template>
-
+<ng-container *ngIf="applyData.type === actionType.table">
+  <object-designer-layout-content-detail-table [applyData]="applyData">
+  </object-designer-layout-content-detail-table>
+</ng-container>
