--- src/renderer/app/views/object_designer/layout/content/operator/table/create/operator.table.create.view.html
@@ -1,50 +1,0 @@
-<nz-tabset>
-  <nz-tab [nzTitle]="titleTemplate">
-    <ng-template #titleTemplate>
-      <i class="fa fa-columns"></i>
-      {{'common.column'|translate}} <span *ngIf="applyColumns.length > 0">({{applyColumns.length}})</span>
-    </ng-template>
-    <div *ngIf="applyColumns.length > 0" style="margin-top: 4px;">
-      <nz-table #tableContainer nzSize="small" [nzFrontPagination]="false"
-                [nzShowPagination]="false" [nzData]="applyColumns">
-        <thead>
-        <tr>
-          <th>{{'common.name'|translate}}</th>
-          <th>{{'common.type'|translate}}</th>
-          <th>{{'common.length'|translate}}</th>
-          <th>{{'common.is_null'|translate}}</th>
-          <th>{{'common.key'|translate}}</th>
-          <th>{{'common.comment'|translate}}</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr *ngFor="let data of tableContainer.data">
-          <td nzEllipsis>
-            <input nz-input [(ngModel)]="data['name']"/>
-          </td>
-          <td>
-            <nz-select [(ngModel)]="data['type']" style="min-width: 120px;">
-              <nz-option nzValue="jack" nzLabel="Jack"></nz-option>
-              <nz-option nzValue="lucy" nzLabel="Lucy"></nz-option>
-              <nz-option nzValue="disabled" nzLabel="Disabled" nzDisabled></nz-option>
-            </nz-select>
-          </td>
-          <td>
-            <nz-input-number [(ngModel)]="data['length']" [nzMin]="1" [nzStep]="1"></nz-input-number>
-          </td>
-          <td>
-            <nz-switch [(ngModel)]="data['isNull']"></nz-switch>
-          </td>
-          <td>
-            <nz-switch [(ngModel)]="data['isPrimaryKey']"></nz-switch>
-          </td>
-          <td>
-            <textarea rows="1" nz-input [(ngModel)]="data['comment']"></textarea>
-          </td>
-        </tr>
-        </tbody>
-      </nz-table>
-    </div>
-  </nz-tab>
-</nz-tabset>
-
