--- src/renderer/app/views/object_designer/layout/header/database/header.database.view.html
@@ -1,7 +1,0 @@
-<ng-container>
-  <button nz-button nzSize="small" nzType="text"
-          nz-tooltip nzTooltipTitle="{{'common.new'|translate}}{{'common.table'|translate}}"
-          (click)="handlerNewTable()">
-    <i nz-icon nzType="plus-circle" class="action-button-success"></i>
-  </button>
-</ng-container>
