--- src/renderer/app/views/object_designer/layout/header/layout.header.style.scss
@@ -1,5 +1,0 @@
-:host ::ng-deep .custom-select {
-  width: 200px;
-  border-right: #f5f6f5 1px solid;
-  margin-right: 10px;
-}
