--- src/renderer/app/views/object_designer/layout/header/layout.header.view.html
@@ -1,16 +1,0 @@
-<nz-select class="custom-select" [(ngModel)]="applyDataSource.current" nzBorderless nzSize="small"
-           (ngModelChange)="handlerDataSourceChange($event)">
-  <nz-option nzCustomContent *ngFor="let dataSource of applyDataSource.list" [nzLabel]="dataSource.alias"
-             [nzValue]="dataSource.alias" [nzDisabled]="!dataSource.status">
-    <nz-avatar nzSize="small" nzSrc="./renderer/assets/icon/source/{{dataSource.type}}.svg">
-    </nz-avatar>
-    {{dataSource.alias}}
-  </nz-option>
-</nz-select>
-<object-designer-layout-header-database [applyData]="applyData"
-                                        *ngIf="isNoneDatabase()">
-</object-designer-layout-header-database>
-<object-designer-layout-header-operator-table-create [applyData]="applyData" [applyColumns]="applyColumns"
-                                                     (emitterColumns)="handlerEmitterColumns($event)"
-                                                     *ngIf="isCreateTable()">
-</object-designer-layout-header-operator-table-create>
