+++ src/renderer/assets/i18n/zh.json
@@ -176,7 +176,8 @@
     "is_null": "不是Null",
     "key": "主键",
     "length": "长度",
-    "beta": "Beta"
+    "beta": "Beta",
+    "sql": "SQL"
   },
   "language": {
     "english": "英语",
